refactor: revamp redis calls
